move internal dtos to be with the queries that use them
