+++ src/SFA.DAS.ProviderRelationships.UnitTests/Application/Queries/GetHealthCheckQueryHandlerTests.cs
@@ -9,8 +9,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using NUnit.Framework;
 using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck.Dtos;
 using SFA.DAS.ProviderRelationships.Data;
-using SFA.DAS.ProviderRelationships.Dtos;
 using SFA.DAS.ProviderRelationships.Mappings;
 using SFA.DAS.ProviderRelationships.Models;
 using SFA.DAS.ProviderRelationships.UnitTests.Builders;
