+++ src/SFA.DAS.ProviderRelationships.Web.UnitTests/Controllers/HealthCheckControllerTests.cs
@@ -10,7 +10,7 @@ using NUnit.Framework;
 using SFA.DAS.ProviderRelationships.Application.Commands.RunHealthCheck;
 
 using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck.Dtos;
 using SFA.DAS.ProviderRelationships.Web.Controllers;
 using SFA.DAS.ProviderRelationships.Web.RouteValues.HealthCheck;
 using SFA.DAS.ProviderRelationships.Web.ViewModels.HealthCheck;
