+++ src/SFA.DAS.ProviderRelationships.Web/ViewModels/AccountProviderLegalEntities/GetAccountProviderLegalEntityViewModel.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity.Dtos;
+using SFA.DAS.ProviderRelationships.DtosShared;
 
 namespace SFA.DAS.ProviderRelationships.Web.ViewModels.AccountProviderLegalEntities
 {
