+++ src/SFA.DAS.ProviderRelationships.Web/ViewModels/AccountProviderLegalEntities/UpdateAccountProviderLegalEntityViewModel.cs
@@ -2,7 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using SFA.DAS.Authorization;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity.Dtos;
+using SFA.DAS.ProviderRelationships.DtosShared;
 
 namespace SFA.DAS.ProviderRelationships.Web.ViewModels.AccountProviderLegalEntities
 {
