+++ src/SFA.DAS.ProviderRelationships.Web/ViewModels/AccountProviders/AlreadyAddedAccountProviderViewModel.cs
@@ -1,5 +1,5 @@
 using System.ComponentModel.DataAnnotations;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.DtosShared;
 
 namespace SFA.DAS.ProviderRelationships.Web.ViewModels.AccountProviders
 {
