+++ src/SFA.DAS.ProviderRelationships.Web/ViewModels/AccountProviders/GetAccountProviderViewModel.cs
@@ -1,4 +1,5 @@
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProvider;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProvider.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Web.ViewModels.AccountProviders
 {
