+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProvider/GetAccountProviderQueryResult.cs
@@ -1,4 +1,4 @@
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProvider.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProvider
 {
