+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProviderLegalEntity/GetAccountProviderLegalEntityQueryResult.cs
@@ -1,4 +1,5 @@
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity.Dtos;
+using SFA.DAS.ProviderRelationships.DtosShared;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity
 {
