+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAccountProviders/Dtos/AccountProviderSummaryDto.cs
@@ -1,4 +1,4 @@
-namespace SFA.DAS.ProviderRelationships.Dtos
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviders.Dtos
 {
     public class AccountProviderSummaryDto
     {
