+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetAddedAccountProvider/GetAddedAccountProviderQueryHandler.cs
@@ -7,7 +7,7 @@ using AutoMapper.QueryableExtensions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using SFA.DAS.ProviderRelationships.Data;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.DtosShared;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetAddedAccountProvider
 {
