+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetHealthCheck/Dtos/HealthCheckDto.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SFA.DAS.ProviderRelationships.Dtos
+namespace SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck.Dtos
 {
     public class HealthCheckDto
     {
