+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetHealthCheck/GetHealthCheckQueryHandler.cs
@@ -6,8 +6,8 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck.Dtos;
 using SFA.DAS.ProviderRelationships.Data;
-using SFA.DAS.ProviderRelationships.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck
 {
