+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetHealthCheck/GetHealthCheckQueryResult.cs
@@ -1,4 +1,4 @@
-﻿using SFA.DAS.ProviderRelationships.Dtos;
+﻿using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck
 {
