+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetProviderToAdd/GetProviderToAddQueryResult.cs
@@ -1,4 +1,4 @@
-﻿using SFA.DAS.ProviderRelationships.Dtos;
+﻿using SFA.DAS.ProviderRelationships.Application.Queries.GetProviderToAdd.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetProviderToAdd
 {
