+++ src/SFA.DAS.ProviderRelationships/Application/Queries/GetUpdatedAccountProviderLegalEntity/GetUpdatedAccountProviderLegalEntityQueryResult.cs
@@ -1,4 +1,4 @@
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetUpdatedAccountProviderLegalEntity.Dtos;
 
 namespace SFA.DAS.ProviderRelationships.Application.Queries.GetUpdatedAccountProviderLegalEntity
 {
