+++ src/SFA.DAS.ProviderRelationships/Mappings/AccountProviderLegalEntityMappings.cs
@@ -1,6 +1,7 @@
 using System.Linq;
 using AutoMapper;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetAccountProviderLegalEntity.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetUpdatedAccountProviderLegalEntity.Dtos;
 using SFA.DAS.ProviderRelationships.Models;
 using SFA.DAS.ProviderRelationships.Types.Dtos;
