+++ src/SFA.DAS.ProviderRelationships/Mappings/HealthCheckMappings.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetHealthCheck.Dtos;
 using SFA.DAS.ProviderRelationships.Models;
 
 namespace SFA.DAS.ProviderRelationships.Mappings
