+++ src/SFA.DAS.ProviderRelationships/Mappings/ProviderMappings.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using SFA.DAS.ProviderRelationships.Dtos;
+using SFA.DAS.ProviderRelationships.Application.Queries.GetProviderToAdd.Dtos;
 using SFA.DAS.ProviderRelationships.Models;
 
 namespace SFA.DAS.ProviderRelationships.Mappings
